Introduce `HrefResolver`
This commit separates the concerns of defining a referrable model in declarative
way, and resolve hyperlinks of that model. Previously the latter task was
implemented by subclasses of `BaseReferrableModel` (that is
`starlette.ReferrableModel`). That created awkward coupling between the models
and the web framework.

This commit instead introduces `HrefResolver` mechanism to bridge the web
framework integration and models.