starlette: ensure correct route chain when converting URL to key
Before this commit, converting URL to a key of given model would succeed as long
as the URL matches to some URL with correct key structure. This is changed by
ensuring that the correct route chain (i.e. mounts + final route with correct
name) is used to match URLs to convert, not just by matching the URL to _some_
route.

This also gives options to more structured way to resolve/cache URL chains, so
the commit contains quite a bit refactoring.